+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
   - 🌟 [validator] 新增 `@mixte/validator` 包
+  - 🌟 [validator] 新增 `isEmail` 方法, 新增 `isEmailReg` 正则导出
 
 ## [v1.11.0], [v1.11.0-beta.1]
   - 📅 2024-04-24